optional interupts
Added the ability to use or not interrupts for all conversions.
streamlined the logic that checks for end of conversion for all functions.
replaced the custom delay function with the freertos one. We could make this optional but when running on freertos, its better to use vTaskDelay to yield the processor.